requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==3.0.4
zipp==3.20.1
requests==2.31.0
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1  # Added Flask-SQLAlchemy
bleach==6.1.0  # Added Bleach